Code refacto + fix default rule removal
